requirements.txt
# Production dependencies for NBA Predictor AI

# Core API
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Admin panel (auth)
bcrypt>=4.0.0
PyJWT>=2.8.0

# ML & Data (MEMORY-OPTIMIZED: removed sklearn, joblib, scipy)
xgboost>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# AI Integration
groq>=0.4.0
tenacity>=8.2.0
duckduckgo-search>=3.0.0

# Config store (Supabase/PostgreSQL)
psycopg2-binary>=2.9.0

# Utilities
requests>=2.31.0
toml>=0.10.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0

# Sports data
sbrscrape>=0.0.10

# Background scheduler
apscheduler>=3.10.0

# HTTP client for keep-alive self-ping
httpx>=0.27.0